update view price from int to decimal

+++ Services/Hotel/HMS.Hol.Dtos/RoomTypeManager/UpdatePriceInHoliday.cs
@@ -13,10 +13,10 @@ namespace HMS.Hol.Dtos.RoomTypeManager
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
 
-        [Range(1, int.MaxValue, ErrorMessage = "Giá phòng/giờ phải lớn hơn 0")]
-        public int PricePerHour { get; set; }
+        [Range(1, float.MaxValue, ErrorMessage = "Giá phòng/giờ phải lớn hơn 0")]
+        public decimal PricePerHour { get; set; }
 
-        [Range(1, int.MaxValue, ErrorMessage = "Giá phòng/giờ phải lớn hơn 0")]
-        public int PricePerNight { get; set; }
+        [Range(1, float.MaxValue, ErrorMessage = "Giá phòng/giờ phải lớn hơn 0")]
+        public decimal PricePerNight { get; set; }
     }
 }
